pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nova-parser"
version = "0.1.0"
description = "書籍からゲームデータを Gemini で抽出する"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai",
    "python-dotenv",
    "pydantic",
    "pydantic-ai",
]

[project.scripts]
nova-parser = "nova_parser.main:main"

[dependency-groups]
dev = [
    "taskipy",
]


[tool.ruff]
target-version = "py314"
exclude = [
    ".venv",
    ".ruff_cache"
]
line-length = 119

[tool.ruff.lint]
select = [
    "F", # Flake8
    "B", # Black
    "I", # isort
    "E", # error
    "W"  # warning
]
ignore = []
fixable = ["ALL"]

# 自動フォーマット設定
[tool.ruff.format]
#quote-style = "double"
#indent-style = "space"
#skip-magic-trailing-comma = false
line-ending = "auto"

[tool.taskipy.tasks]
ruff = "uvx ruff check --fix --unsafe-fixes src/ && uvx ruff format src/"

